+++ src/lib_dal_node/constants.mli
@@ -43,6 +43,10 @@ val cache_size : int
     associate commitments with slot ids at a given level. *)
 val slot_id_cache_size : int
 
+(** [statuses_cache_size] is the size (in number of slots) of the cache
+    to associate slot ids to slot status. *)
+val statuses_cache_size : int
+
 (** The frequency at which we sample the time spent in shards crypto
     verification. *)
 val shards_verification_sampling_frequency : int