+++ src/lib_dal_node/store.ml
@@ -25,6 +25,17 @@
 
 module Profiler = (val Profiler.wrap Dal_profiler.dal_profiler)
 
+(** FIFO-keyed map with slot_id as keys. *)
+module Slot_map =
+  Aches.Vache.Map (Aches.Vache.FIFO_Precise) (Aches.Vache.Strong)
+    (struct
+      type t = Types.Slot_id.t
+
+      let equal = Types.Slot_id.equal
+
+      let hash = Types.Slot_id.hash
+    end)
+
 module Version = struct
   type t = int
 
@@ -46,8 +57,9 @@ module Version = struct
      - 0: came with octez release v20; used Irmin for storing slots
      - 1: removed Irmin dependency; added slot and status stores; changed layout of shard
        store by indexing on slot ids instead of commitments
-     - 2: switch the KVS skip list store for a sqlite3 one. *)
-  let current_version = 2
+     - 2: switch the KVS skip list store for a sqlite3 one.
+     - 3: remove status store, keep cache. *)
+  let current_version = 3
 
   type error += Could_not_read_data_dir_version of string
 
@@ -119,11 +131,11 @@ module Version = struct
       (fun () -> Data_encoding.Json.destruct encoding json |> return)
       (fun _ -> tzfail (Could_not_read_data_dir_version file_path))
 
-  let write_version_file ~base_dir =
+  let write_version_file ~version ~base_dir =
     let version_file = version_file_path ~base_dir in
     Lwt_utils_unix.Json.write_file
       version_file
-      (Data_encoding.Json.construct encoding current_version)
+      (Data_encoding.Json.construct encoding version)
     |> trace (Could_not_write_version_file version_file)
 end
 
@@ -133,8 +145,6 @@ module Stores_dirs = struct
   let shard = "shard_store"
 
   let slot = "slot_store"
-
-  let status = "status_store"
 
   let skip_list_cells = "skip_list_store"
 end
@@ -320,18 +330,6 @@ module Shards_disk = struct
 end
 
 module Shards_cache = struct
-  (** Underlying FIFO-keyed map from slot_id -> map from shard index to share. This is
-      used as the alternative to the on-disk storage [Shards_disk] for shards. *)
-  module Slot_map =
-    Aches.Vache.Map (Aches.Vache.FIFO_Precise) (Aches.Vache.Strong)
-      (struct
-        type t = Types.Slot_id.t
-
-        let equal = Types.Slot_id.equal
-
-        let hash = Types.Slot_id.hash
-      end)
-
   module Int_map = Map.Make (Int)
 
   type t = Cryptobox.share Int_map.t Slot_map.t
@@ -663,91 +661,39 @@ module Traps = struct
           []
 end
 
-module Statuses = struct
-  type t = (int32, int, Types.header_status) KVS.t
+module Statuses_cache = struct
+  type t = Types.header_status Slot_map.t
 
-  let file_layout ~root_dir slot_level =
-    (* The number of entries per file is the number of slots. We put
-       here the max value (4096) because we don't have a cryptobox
-       at hand to get the number_of_slots parameter. *)
-    let number_of_keys_per_file = 4096 in
-    let level_string = Format.asprintf "%ld" slot_level in
-    let filepath = Filename.concat root_dir level_string in
-    Key_value_store.layout
-      ~encoding:Types.header_status_encoding
-      ~filepath
-      ~eq:Stdlib.( = )
-      ~index_of:Fun.id
-      ~number_of_keys_per_file
-      ()
-
-  let init node_store_dir status_store_dir =
-    let root_dir = Filename.concat node_store_dir status_store_dir in
-    KVS.init ~lru_size:Constants.status_store_lru_size ~root_dir
+  let init = Slot_map.create
 
   let add_status t status (slot_id : Types.slot_id) =
-    let open Lwt_result_syntax in
-    let* () =
-      KVS.write_value
-        ~override:true
-        t
-        file_layout
-        slot_id.slot_level
-        slot_id.slot_index
-        status
-      |> Errors.other_lwt_result
-    in
-    let*! () =
-      Event.emit_stored_slot_status
-        ~published_level:slot_id.slot_level
-        ~slot_index:slot_id.slot_index
-        ~status
-    in
-    return_unit
+    Slot_map.replace t slot_id status
 
-  let find_status t (slot_id : Types.slot_id) =
-    let open Lwt_result_syntax in
-    let*! res =
-      KVS.read_value t file_layout slot_id.slot_level slot_id.slot_index
-    in
-    match res with
-    | Ok status -> return status
-    | Error [KVS.Missing_stored_kvs_data _] -> fail Errors.not_found
-    | Error err ->
-        let data_kind = Types.Store.Header_status in
-        fail @@ Errors.decoding_failed data_kind err
+  let get_slot_status = Slot_map.find_opt
 
   let update_slot_headers_attestation ~published_level ~number_of_slots t
       attested =
-    let open Lwt_result_syntax in
-    List.iter_es
+    List.iter
       (fun slot_index ->
         let index = Types.Slot_id.{slot_level = published_level; slot_index} in
         if attested slot_index then (
           Dal_metrics.slot_attested ~set:true slot_index ;
-          add_status t `Attested index |> Errors.to_tzresult)
+          add_status t `Attested index)
         else
-          let* old_data_opt =
-            find_status t index |> Errors.to_option_tzresult
-          in
+          let old_data_opt = get_slot_status t index in
           Dal_metrics.slot_attested ~set:false slot_index ;
-          if Option.is_some old_data_opt then
-            add_status t `Unattested index |> Errors.to_tzresult
+          if Option.is_some old_data_opt then add_status t `Unattested index
           else
             (* There is no header that has been included in a block
                and selected for this index. So, the slot cannot be
                attested or unattested. *)
-            return_unit)
+            ())
       (0 -- (number_of_slots - 1))
 
   let update_selected_slot_headers_statuses ~block_level ~attestation_lag
       ~number_of_slots attested t =
     let published_level = Int32.(sub block_level (of_int attestation_lag)) in
     update_slot_headers_attestation ~published_level ~number_of_slots t attested
-
-  let get_slot_status ~slot_id t = find_status t slot_id
-
-  let remove_level_status ~level t = KVS.remove_file t file_layout level
 end
 
 module Commitment_indexed_cache =
@@ -848,7 +794,7 @@ end
 
 (** Store context *)
 type t = {
-  slot_header_statuses : Statuses.t;
+  statuses_cache : Statuses_cache.t;
   shards : Shards.t;
   slots : Slots.t;
   traps : Traps.t;
@@ -877,7 +823,7 @@ let shards {shards; _} = shards
 
 let skip_list_cells t = t.skip_list_cells_store
 
-let slot_header_statuses {slot_header_statuses; _} = slot_header_statuses
+let statuses_cache {statuses_cache; _} = statuses_cache
 
 let slots {slots; _} = slots
 
@@ -944,12 +890,35 @@ let cache_entry node_store commitment slot shares shard_proofs =
     commitment
     (slot, shares, shard_proofs)
 
+let upgrade_from_v2_to_v3 ~base_dir =
+  let open Lwt_result_syntax in
+  let*! () =
+    Event.emit_store_upgrade_start
+      ~old_version:(Version.make 2)
+      ~new_version:(Version.make 3)
+  in
+  let file_path = Filename.concat base_dir "status_store" in
+  let*! exists = Lwt_unix.file_exists file_path in
+  let*! () =
+    if exists then Lwt_utils_unix.remove_dir file_path else Lwt.return_unit
+  in
+  Version.write_version_file ~base_dir ~version:3
+
+(* [upgradable old_version new_version] returns an upgrade function if
+   the store is upgradable from [old_version] to [new_version]. Otherwise it
+   returns [None]. *)
+let upgradable old_version new_version :
+    (base_dir:string -> unit tzresult Lwt.t) option =
+  match (old_version, new_version) with
+  | 2, 3 -> Some upgrade_from_v2_to_v3
+  | _ -> None
+
 (* Checks the version of the store with the respect to the current
    version. Returns [None] if the store does not need an upgrade and [Some
    upgrade] if the store is upgradable, where [upgrade] is a function that can
    be used to upgrade the store. It returns an error if the version is
    incompatible with the current one. *)
-let check_version_and_may_upgrade base_dir =
+let rec check_version_and_may_upgrade base_dir =
   let open Lwt_result_syntax in
   let file_path = Version.version_file_path ~base_dir in
   let*! exists = Lwt_unix.file_exists file_path in
@@ -958,26 +927,37 @@ let check_version_and_may_upgrade base_dir =
     else
       (* In the absence of a version file, we use an heuristic to determine the
          version. *)
-      let*! exists = Lwt_unix.file_exists (Filename.concat base_dir "index") in
-      return
-      @@ if exists then Version.make 0 else Version.make Version.current_version
+      let*! index = Lwt_unix.file_exists (Filename.concat base_dir "index") in
+      if index then return (Version.make 0)
+      else
+        let*! status =
+          Lwt_unix.file_exists (Filename.concat base_dir "status_store")
+        in
+        if status then return (Version.make 2)
+        else return (Version.make Version.current_version)
   in
   if Version.(equal version current_version) then return_unit
   else
-    tzfail
-      (Version.Invalid_data_dir_version
-         {actual = version; expected = Version.current_version})
+    match upgradable version Version.current_version with
+    | Some upgrade ->
+        let* () = upgrade ~base_dir in
+        (* Now that we upgraded, check version again *)
+        check_version_and_may_upgrade base_dir
+    | None ->
+        tzfail
+          (Version.Invalid_data_dir_version
+             {actual = version; expected = Version.current_version})
 
 let init config profile_ctxt proto_parameters =
   let open Lwt_result_syntax in
   let base_dir = Configuration_file.store_path config in
   let* () = check_version_and_may_upgrade base_dir in
-  let* slot_header_statuses = Statuses.init base_dir Stores_dirs.status in
+  let statuses_cache = Statuses_cache.init Constants.statuses_cache_size in
   let* shards =
     Shards.init ~profile_ctxt ~proto_parameters base_dir Stores_dirs.shard
   in
   let* slots = Slots.init base_dir Stores_dirs.slot in
-  let* () = Version.write_version_file ~base_dir in
+  let* () = Version.(write_version_file ~base_dir ~version:current_version) in
   let traps = Traps.create ~capacity:Constants.traps_cache_size in
   let* chain_id = Chain_id.init ~root_dir:base_dir in
   let* last_processed_level = Last_processed_level.init ~root_dir:base_dir in
@@ -989,7 +969,7 @@ let init config profile_ctxt proto_parameters =
       shards;
       slots;
       traps;
-      slot_header_statuses;
+      statuses_cache;
       cache = Commitment_indexed_cache.create Constants.cache_size;
       finalized_commitments =
         Slot_id_cache.create ~capacity:Constants.slot_id_cache_size;
@@ -1001,10 +981,10 @@ let init config profile_ctxt proto_parameters =
 
 let add_slot_headers ~number_of_slots ~block_level slot_headers t =
   let module SI = Set.Make (Int) in
-  let open Lwt_result_syntax in
-  let slot_header_statuses = t.slot_header_statuses in
+  let open Lwt_syntax in
+  let statuses_cache = t.statuses_cache in
   let* waiting =
-    List.fold_left_es
+    List.fold_left_s
       (fun waiting slot_header ->
         let Dal_plugin.{slot_index; commitment = _; published_level} =
           slot_header
@@ -1012,9 +992,8 @@ let add_slot_headers ~number_of_slots ~block_level slot_headers t =
         (* This invariant should hold. *)
         assert (Int32.equal published_level block_level) ;
         let index = Types.Slot_id.{slot_level = published_level; slot_index} in
-        let* () =
-          Statuses.add_status slot_header_statuses `Waiting_attestation index
-          |> Errors.to_tzresult
+        let () =
+          Statuses_cache.add_status statuses_cache `Waiting_attestation index
         in
         Slot_id_cache.add ~number_of_slots t.finalized_commitments slot_header ;
         return (SI.add slot_index waiting))